# openapi.yaml
openapi: 3.1.0
info:
  title: FitBuddy API
  version: "1.0.0"
  description: |
    Auth, Challenge, Exercise/Plan/Schedule, Records, GPT 루틴, 세트 관리 등 엔드포인트 문서.
    이 파일이 존재하면 swagger.js가 이 스펙을 사용합니다.
servers:
  - url: https://fitbuddy-server-p6r6.onrender.com
    description: Production
  - url: http://localhost:3000
    description: Local
...
components:
  schemas:
    ApiMessage:
      type: object
      properties:
        message:
          type: string
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    SendOtpRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    VerifyOtpRequest:
      type: object
      required: [email, otp]
      properties:
        email: { type: string, format: email }
        otp: { type: string, description: "6자리 숫자 문자열" }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    UpdateProfileRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        name: { type: string }
        age_group: { type: string }
        gender: { type: string }
        height: { type: number }
        weight: { type: number }
        diseases:
          type: array
          items: { type: string }
        workout_level: { type: string }
        preferred_workouts:
          type: array
          items: { type: string }
        equipment:
          type: array
          items: { type: string }
    UserProfile:
      type: object
      properties:
        email: { type: string, format: email }
        name: { type: string }
        age_group: { type: string }
        gender: { type: string }
        height: { type: number }
        weight: { type: number }
        diseases:
          type: array
          items: { type: string }
        workout_level: { type: string }
        preferred_workouts:
          type: array
          items: { type: string }
        equipment:
          type: array
          items: { type: string }
    UserInfoById:
      type: object
      properties:
        name: { type: string }
        age_group: { type: string }
        gender: { type: string }
        height: { type: number }
        weight: { type: number }
        disease: { type: string }
        exercise_level: { type: string }
        preferred_exercises:
          type: array
          items: { type: string }
        exercise_equipment:
          type: array
          items: { type: string }
...
paths:
  /signup:
    post:
      tags: [Auth]
      summary: 회원가입 정보 임시 저장
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ApiMessage' }}}}
  /send-otp:
    post:
      tags: [Auth]
      summary: 이메일로 OTP 전송
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendOtpRequest' }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ApiMessage' }}}}
        "400": { description: 잘못된 이메일 }
        "500": { description: 전송 실패 }
  /verify-otp:
    post:
      tags: [Auth]
      summary: OTP 검증 후 DB 저장
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOtpRequest' }
      responses:
        "200": { description: 가입/이미 인증 완료, content: { application/json: { schema: { $ref: '#/components/schemas/ApiMessage' }}}}
        "400": { description: 불일치/만료 또는 사전정보 없음 }
        "500": { description: DB 오류 }
  /update-profile:
    post:
      tags: [Profile]
      summary: 사용자 프로필 업데이트(이메일 기준)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        "200": { description: 저장 완료, content: { application/json: { schema: { $ref: '#/components/schemas/ApiMessage' }}}}
        "404": { description: 사용자 없음 }
        "500": { description: DB 실패 }
  /login:
    post:
      tags: [Auth]
      summary: 로그인(이메일/패스워드)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        "200": { description: 로그인 성공, content: { application/json: { schema: { type: object, properties: { message: {type: string}, user: { type: object }}}}}}
        "401": { description: 불일치 }
        "500": { description: 서버 오류 }
  /get-profile:
    get:
      tags: [Profile]
      summary: 이메일로 프로필 조회
      parameters:
        - in: query
          name: email
          required: true
          schema: { type: string, format: email }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserProfile' }}}}
        "404": { description: 없음 }
        "500": { description: 서버 오류 }
  /api/user-info/{userId}:
    get:
      tags: [User]
      summary: userId로 사용자 정보 조회
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserInfoById' }}}}
        "404": { description: 없음 }
        "500": { description: 서버 오류 }
...